refactor: 중복타입 공통으로 분리
