+++ src/app/(with-header)/(protected)/_types/index.ts
@@ -12,3 +12,10 @@ export interface BookType {
   thumbnailUrl: string
   review: ReviewType
 }
+
+export interface ProfileType {
+  nickName: string
+  email: string
+  created_at: string
+  updated_at: string
+}
