+++ src/app/(with-header)/(protected)/my-page/_types/index.ts
@@ -50,18 +50,3 @@ export interface TagBookType {
   count: number
   percent: number
 }
-
-export interface ReviewType {
-  endDate: string
-  memo: string
-  rating: number
-  tags: [string]
-}
-
-export interface RecentBookType {
-  id: string
-  title: string
-  author: string
-  thumbnailUrl: string
-  review: ReviewType
-}
