--- src/app/(with-header)/(protected)/profile/_types/index.ts
@@ -1,6 +1,0 @@
-export interface ProfileType {
-  nickName: string
-  email: string
-  created_at: string
-  updated_at: string
-}
